bots/tictocdoc/requirements.txt
python-telegram-bot>=20.4,<21
yt-dlp>=2023.11.16
